Adding support for invertedquantitative scale
