+++ src/constants.js
@@ -26,4 +26,33 @@ export const nonFilters = [
   "bill",
   "link"
 ];
+
 export const nonFilterPrefix = 'na_'
+
+export const qualKeys = [
+  "no",
+  "yes_low",
+  "yes_high",
+  "proposed_low",
+  "proposed_high",
+  "no-no",
+  "no-yes_low",
+  "no-yes_high",
+  "yes_low-no",
+  "yes_low-yes_low",
+  "yes_low-yes_high",
+  "yes_high-no",
+  "yes-high_yes-low",
+  "yes-high_yes-high"
+];
+
+export const quantitativeColors = [
+  "#67001f",
+  "#b2182b",
+  "#d6604d",
+  "#f4a582",
+  "#92c5de",
+  "#4393c3",
+  "#2166ac",
+  "#053061"
+];
